requirements.txt
webvtt-py
anthropic
python-dotenv